+++ apps/studio.giselles.ai/app/(main)/settings/team/vector-stores/page.tsx
@@ -1,5 +1,4 @@
 import { ExternalLink } from "lucide-react";
-import { multiEmbeddingFlag } from "@/flags";
 import { getGitHubIdentityState } from "@/services/accounts";
 import {
 	deleteRepositoryIndex,
@@ -38,12 +37,10 @@ export default async function TeamVectorStorePage() {
 		return <GitHubAppInstallRequiredCard />;
 	}
 
-	const [installationsWithRepos, repositoryIndexes, multiEmbedding] =
-		await Promise.all([
-			getInstallationsWithRepos(),
-			getGitHubRepositoryIndexes(),
-			multiEmbeddingFlag(),
-		]);
+	const [installationsWithRepos, repositoryIndexes] = await Promise.all([
+		getInstallationsWithRepos(),
+		getGitHubRepositoryIndexes(),
+	]);
 
 	return (
 		<div className="flex flex-col gap-[24px]">
@@ -69,7 +66,6 @@ export default async function TeamVectorStorePage() {
 					<RepositoryRegistrationDialog
 						installationsWithRepos={installationsWithRepos}
 						registerRepositoryIndexAction={registerRepositoryIndex}
-						multiEmbedding={multiEmbedding}
 					/>
 				</div>
 			</div>
@@ -79,7 +75,6 @@ export default async function TeamVectorStorePage() {
 				deleteRepositoryIndexAction={deleteRepositoryIndex}
 				triggerManualIngestAction={triggerManualIngest}
 				updateRepositoryIndexAction={updateRepositoryIndex}
-				multiEmbedding={multiEmbedding}
 			/>
 		</div>
 	);
