+++ apps/studio.giselles.ai/app/(main)/settings/team/vector-stores/repository-list.tsx
@@ -3,7 +3,6 @@ import type { GitHubRepositoryContentType } from "@/drizzle";
 import type { RepositoryWithStatuses } from "@/lib/vector-stores/github";
 import type { GitHubRepositoryIndexId } from "@/packages/types";
 import { RepositoryItem } from "./repository-item";
-import { RepositoryItem as RepositoryItemV2 } from "./repository-item-v2";
 
 type RepositoryListProps = {
 	repositories: RepositoryWithStatuses[];
@@ -21,7 +20,6 @@ type RepositoryListProps = {
 		}[],
 		embeddingProfileIds?: number[],
 	) => Promise<{ success: boolean; error?: string }>;
-	multiEmbedding?: boolean;
 };
 
 export function RepositoryList({
@@ -29,7 +27,6 @@ export function RepositoryList({
 	deleteRepositoryIndexAction,
 	triggerManualIngestAction,
 	updateRepositoryIndexAction,
-	multiEmbedding = false,
 }: RepositoryListProps) {
 	return (
 		<div className="flex flex-col gap-y-[16px]">
@@ -48,21 +45,15 @@ export function RepositoryList({
 
 				{repositories.length > 0 ? (
 					<div className="space-y-4">
-						{repositories.map((repo) => {
-							const Component = multiEmbedding
-								? RepositoryItemV2
-								: RepositoryItem;
-							return (
-								<Component
-									key={repo.repositoryIndex.id}
-									repositoryData={repo}
-									deleteRepositoryIndexAction={deleteRepositoryIndexAction}
-									triggerManualIngestAction={triggerManualIngestAction}
-									updateRepositoryIndexAction={updateRepositoryIndexAction}
-									multiEmbedding={multiEmbedding}
-								/>
-							);
-						})}
+						{repositories.map((repo) => (
+							<RepositoryItem
+								key={repo.repositoryIndex.id}
+								repositoryData={repo}
+								deleteRepositoryIndexAction={deleteRepositoryIndexAction}
+								triggerManualIngestAction={triggerManualIngestAction}
+								updateRepositoryIndexAction={updateRepositoryIndexAction}
+							/>
+						))}
 					</div>
 				) : (
 					<EmptyRepositoryCard />
