+++ apps/studio.giselles.ai/app/workspaces/[workspaceId]/layout.tsx
@@ -7,7 +7,6 @@ import {
 	aiGatewayFlag,
 	experimental_storageFlag,
 	layoutV3Flag,
-	multiEmbeddingFlag,
 	resumableGenerationFlag,
 	runV3Flag,
 	stageFlag,
@@ -51,7 +50,6 @@ export default async function Layout({
 	const layoutV3 = await layoutV3Flag();
 	const experimental_storage = await experimental_storageFlag();
 	const stage = await stageFlag();
-	const multiEmbedding = await multiEmbeddingFlag();
 	const aiGateway = await aiGatewayFlag();
 	const resumableGeneration = await resumableGenerationFlag();
 
@@ -87,7 +85,6 @@ export default async function Layout({
 				layoutV3,
 				experimental_storage,
 				stage,
-				multiEmbedding,
 				aiGateway,
 				resumableGeneration,
 			}}
