+++ packages/giselle/src/react/workspace/provider.tsx
@@ -68,7 +68,6 @@ export function WorkspaceProvider({
 				layoutV3: featureFlag?.layoutV3 ?? false,
 				experimental_storage: featureFlag?.experimental_storage ?? false,
 				stage: featureFlag?.stage ?? false,
-				multiEmbedding: featureFlag?.multiEmbedding ?? false,
 				aiGateway: featureFlag?.aiGateway ?? false,
 				resumableGeneration: featureFlag?.resumableGeneration ?? false,
 			}}
